Move `set_up_fabric` back to `operations` to avoid additional fabric import in cmdline.
Should be moved to those modules which really import operations propably.
