+++ src/bundlewrap/operations.py
@@ -15,6 +15,17 @@ from .utils.text import mark_for_translation as _, randstr
 from .utils.ui import LineBuffer
 
 
+def set_up_fabric():
+    """
+    Setup fabric.
+    """
+    env.use_ssh_config = True
+    env.warn_only = True
+    # silence fabric
+    for key in output:
+        output[key] = False
+
+
 class FabricOutput(object):
     def __init__(self, silent=False):
         self.silent = silent
